Add file body handling.
CakePHP's response does most of the hard work in preparing the headers
that are important. I had to hack around to read the file out. I'll have
to make a pull request to cakephp to add a way to read the file in
a non-evil way.

Emitting files will require more work in the Server but that is for
another day.